refactor: Add the mecanism to switch form the legacy to the new linter
Use the new linter/config/inline_config if the config object contain the 'rules' keywords
