+++ lib/index.js
@@ -1,4 +1,5 @@
 const Linter = require("./linter");
+const LegacyLinter = require("./legacy/linter");
 
 /* istanbul ignore next */
 Object.values =
@@ -14,18 +15,22 @@ Object.values =
  * The linthtml namespace.
  * @namespace
  */
-var linthtml = function() {
-  var linter = linthtml.defaultLinter;
+var linthtml = function(html, config) {
 
+  if (config && config.hasOwnProperty('rules')) {
+    const linter = new Linter(config);
+    return linter.linth(html);
+  }
+  const linter = new LegacyLinter();
   return linter.lint.apply(linter, arguments);
 };
 
 module.exports = linthtml;
 
 linthtml.Linter = Linter;
+linthtml.LegacyLinter = LegacyLinter;
 linthtml.rules = require("./rules");
 linthtml.messages = require("./messages");
-linthtml.defaultLinter = new Linter(linthtml.rules);
 
 // linthtml.use = function(plugins) {
 //   plugins.forEach(function(plugin) {
