+++ lib/legacy/config.js
@@ -1,4 +1,4 @@
-const { isBool } = require("./validate_option");
+const { isBool } = require("../validate_option");
 const pull = require('lodash.pull');
 /**
  * The config object stores all possible rules and options and manages
