+++ lib/legacy/inline_config.js
@@ -1,6 +1,6 @@
-const knife = require("./knife");
-const presets = require("./presets");
-const Issue = require("./issue");
+const knife = require("../knife");
+const presets = require("../presets");
+const Issue = require("../issue");
 
 // Private vars,
 var index = 0; // index used for making sure configs are sent in order
