+++ lib/legacy/linter.js
@@ -1,10 +1,10 @@
-const Parser = require("./parser");
-const knife = require("./knife");
-const presets = require("./presets");
-const Config = require("./legacy/config");
-const InlineConfig = require("./legacy/inline_config");
-const { flatten } = require("./utils/arrays");
-
+const Parser = require("../parser");
+const knife = require("../knife");
+const presets = require("../presets");
+const Config = require("./config");
+const InlineConfig = require("./inline_config");
+const { flatten } = require("../utils/arrays");
+const rules = require('../rules');
 /**
  * Apply the raw-ignore-regex option.
  * Return the modified html, and a function that recovers line/column
@@ -33,7 +33,7 @@ function getOptions(args) {
 }
 
 class Linter {
-  constructor(rules, options) {
+  constructor(_, options) {
     this.rules = new Config(rules, options);
     this.parser = new Parser();
     this.inlineConfig = new InlineConfig(this.rules);
