+++ test/unit/inline-configuration.js
@@ -1,4 +1,4 @@
-const InlineConfig = require("../../lib/inline_config_legacy");
+const InlineConfig = require("../../lib/legacy/inline_config");
 const { expect } = require("chai");
 
 const linthtml = require("../../lib");
